updated update_post_partial and delete_post with simillar changes to this previous commit

+++ routers/posts.py
@@ -99,11 +99,18 @@ async def update_post_full(post_id: int, post_data:PostCreate, current_user:Curr
 
 # Update on a post can be title or content, one can change without the other having to be override by default values
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data:PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int, post_data:PostUpdate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result_post_id =    await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result_post_id.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Unauthorize to edit this post"
+        )
 
     update_data = post_data.model_dump(exclude_unset=True)  
     for field, value in update_data.items():
@@ -118,13 +125,17 @@ async def update_post_partial(post_id: int, post_data:PostUpdate, db: Annotated[
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT )
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result_post_id =    await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result_post_id.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
-
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Unauthorize to edit this post"
+        )
     await db.delete(post)
     await db.commit()
